done crud operator

+++ api/routes/admins.py
@@ -55,11 +55,17 @@ async def create_admin(data_in: schemas.admin.AdminCreate, db: Session = Depends
 
 @router.patch("/{target_user_id}", response_model=schemas.admin.PublicAdmin, dependencies=[Depends(is_super_user)])
 async def update_admin(target_user_id: int, data_in: schemas.admin.AdminPartialIn, db: Session = Depends(get_db)):
+    admin = await AdminLogic.get_by_user_id(db, target_user_id)
+    if not admin:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Admin not found")
+    await AdminLogic.update(db, admin.id, data_in.model_dump(exclude_none=True))
     await update_user(target_user_id, schemas.user.ModifyUserByAdmin(**data_in.model_dump(exclude_none=True)), db)
-    admin = await AdminLogic.update_by_user_id(db, target_user_id, data_in.model_dump(exclude_none=True))
     return admin
 
 
 @router.delete("/{target_user_id}", dependencies=[Depends(is_super_user)])
 async def remove_admin(target_user_id: int, db: Session = Depends(get_db)):
+    admin = await AdminLogic.get_by_user_id(db, target_user_id)
+    if not admin:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Admin not found")
     await delete_user(target_user_id, db)
